+++ emgimuservice/src/main/java/org/sralab/emgimu/logging/EmgLogFetchJobService.java
@@ -152,7 +152,7 @@ public class EmgLogFetchJobService extends JobService implements EmgImuManagerCa
     }
 
     @Override
-    public void onEmgBuffReceived(BluetoothDevice device, int count, int[][] data) {
+    public void onEmgBuffReceived(BluetoothDevice device, int count, double[][] data) {
     }
 
     @Override
